# Server/root/swagger.yaml
openapi: 3.0.0
info:
  title: Unidomus API
  version: 2.0.0
  description: API for the Unidomus application, a platform for real estate listings and roommate matching.
paths:
...
  /api/listings:
    get:
      summary: Retrieve listings based on query parameters
      description: Retrieve listings based on price range, typology, city, and floor area range.
      tags:
        - Listings
      parameters:
        - in: query
          name: priceMin
          schema:
            type: number
          description: Minimum price filter
        - in: query
          name: priceMax
          schema:
            type: number
          description: Maximum price filter
        - in: query
          name: typology
          schema:
            type: string
          description: Typology filter
        - in: query
          name: city
          schema:
            type: string
          description: City filter
        - in: query
          name: floorAreaMin
          schema:
            type: number
          description: Minimum floor area filter
        - in: query
          name: floorAreaMax
          schema:
            type: number
          description: Maximum floor area filter
      responses:
        "200":
          description: A list of listings matching the query parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  listings:
                    type: array
                    items:
                      $ref: "#/components/schemas/Listing"
  /api/listings/{id}:
    get:
      summary: Retrieve a listing by ID
      description: Retrieve a listing by its unique identifier.
      tags:
        - Listings
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the listing to retrieve
      responses:
        "200":
          description: Listing retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          description: Listing not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Listing not found
        "500":
          description: Error retrieving listing
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error retrieving listing
                  error:
                    type: string
    delete:
      description: Deletes a listing by its ID. The publisher of the listing is notified
        upon deletion.
      parameters:
      - description: The ID of the listing to delete
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
          description: Listing deleted successfully
        '400':
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
          description: Listing not found or Publisher not found
        '500':
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                  message:
                    type: string
                type: object
          description: Internal server error
      summary: Delete listing by ID
      tags:
      - Listings
    put:
      description: Updates the details of a listing by its ID. This includes updating
        fields such as price, typology, address, description, and more.
      parameters:
      - description: The ID of the listing to update
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                address:
                  properties:
                    cap:
                      type: string
                    city:
                      type: string
                    country:
                      type: string
                    houseNum:
                      type: string
                    province:
                      type: string
                    street:
                      type: string
                  type: object
                availability:
                  type: string
                description:
                  type: string
                floorArea:
                  type: number
                photos:
                  items:
                    type: string
                  type: array
                price:
                  type: number
                publisherID:
                  type: string
                tenantsID:
                  items:
                    type: string
                  type: array
                typology:
                  type: string
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  listing:
                    $ref: '#/components/schemas/Listing'
                type: object
          description: Listing updated successfully
        '400':
          content:
            application/json:
              schema:
                properties:
                  errors:
                    items:
                      properties:
                        field:
                          type: string
                        message:
                          type: string
                      type: object
                    type: array
                  message:
                    type: string
                type: object
          description: Validation errors
        '404':
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
          description: Listing not found
        '500':
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                  message:
                    type: string
                type: object
          description: Internal server error
      summary: Update listing by ID
      tags:
      - Listings
...
components:
  schemas:
...
    Listing:
      type: object
      properties:
        address:
          type: object
          properties:
            street:
              type: string
              minLength: 3
              maxLength: 50
            city:
              type: string
              minLength: 3
              maxLength: 50
            cap:
              type: string
              minLength: 5
              maxLength: 5
              pattern: '^\d{5}$'
            houseNum:
              type: string
              minLength: 1
              maxLength: 5
            province:
              type: string
              minLength: 2
              maxLength: 2
            country:
              type: string
              minLength: 3
              maxLength: 50
        photos:
          type: array
          items:
            type: string
          maxItems: 10
        publisherID:
          type: string
          format: uuid
        tenantsID:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 12
        typology:
          type: string
          maxLength: 30
        description:
          type: string
          maxLength: 1000
        price:
          type: number
          minimum: 10
          maximum: 10000
        floorArea:
          type: number
          minimum: 1
          maximum: 10000
        availability:
          type: string
        publicationDate:
          type: string
          format: date-time
...